requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
anyio==4.9.0
asgiref==3.8.1
autopep8==1.6.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dashboard==0.0.6
databases==0.9.0
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==4.2.10
django-bootstrap5==25.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==3.11.0
django-crispy-forms==2.3
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-storages==1.14.5
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
orm==0.2.0.dev1
pillow==11.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pycodestyle==2.8.0
pycparser==2.22
PyJWT==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2021.3
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.39
sqlparse==0.4.2
starlette==0.46.1
tablib==3.8.0
toml==0.10.2
typesystem==0.3.0.dev0
typing_extensions==4.12.2
tzdata==2021.5
urllib3==2.3.0
uvicorn==0.34.0
gunicorn
